hypercode-core/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.10.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
coverage==7.4.0
httpx==0.26.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk==1.40.0
structlog==24.1.0
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
fakeredis==2.19.0
sse-starlette==1.6.1
